fix build with cmake and rpm

+++ src/xbatd/xbatd.spec
@@ -1,5 +1,9 @@
 # rpmbuild --target x86_64 -bb xbatd.spec
 
+# Disable automatic dependency detection for better control
+%global _enable_debug_package 0
+%global debug_package %{nil}
+%global __os_install_post /usr/lib/rpm/brp-compress %{nil}
 
 Summary: xbat daemon
 Name: xbatd
@@ -8,14 +12,20 @@ Release: %{RELEASE}%{?dist}
 License: MIT
 URL: megware.com
 Packager: Nico Tippmann
-BuildRoot: /root/rpmbuild/
+BuildRoot: %{_tmppath}/%{name}-%{version}-%{release}-root
 Source0: xbatd-%{VERSION}.tar.gz
-BuildRoot:	%{buildroot}
 AutoReqProv: no
+
+# Build requirements
+BuildRequires: cmake >= 3.12
+BuildRequires: gcc-c++
+BuildRequires: make
+
+# Runtime requirements
 Requires: boost-log
 Requires: sysstat
 Requires: libcurl
-%global debug_package %{nil}
+
 %description
 xbat daemon
 
@@ -23,7 +33,6 @@ xbat daemon
 %setup
 
 %build
-# directories already created during likwid installation
 %define BASE /usr/local/share/xbatd/
 %define LIB %{BASE}/lib
 %define LIB64 %{BASE}/lib64
@@ -43,32 +52,47 @@ cp -r /c-questdb-client/build/libquestdb_client.* %{LIB}
 cp -r /usr/lib64/libnvidia-ml.* %{LIB64}
 ln -s %{LIB64}/libnvidia-ml.so.1 %{LIB64}/libnvidia-ml.so
 cp -r /opt/rocm/lib/libamd_smi.* %{LIB}
-rm -rf build && mkdir build && cd build
-cmake .. \
-  -DCMAKE_CXX_FLAGS="-I/opt/rocm/include -I/usr/local/cuda/include" \
-  -DCMAKE_EXE_LINKER_FLAGS="-L/opt/rocm/lib -L/usr/lib64" \
-  -DBUILD_TESTS=OFF \
-  -DBUILD_BENCHMARK=OFF \
-  -DCMAKE_BUILD_TYPE=Release
+# LIKWID is already installed at %{LIB} location, no need to copy
 
-make -j$(nproc)
-#make -e LIB_PATH=%{LIB} LIB64_PATH=%{LIB64} INCLUDE_PATH=%{INCLUDE} -j
+# Clean any existing build directory
+rm -rf build
+
+cmake -B build -S . \
+  -DCMAKE_CXX_FLAGS="-I/opt/rocm/include -I/usr/local/cuda/include" \
+  -DCMAKE_EXE_LINKER_FLAGS="-L/opt/rocm/lib -L/usr/lib64 -L%{LIB} -L%{LIB64}" \
+  -DCMAKE_BUILD_TYPE=Release \
+  -DCMAKE_INSTALL_PREFIX=/usr/local
+
+cmake --build build --parallel %{?_smp_mflags}
 
 %install 
 mkdir -p %{BUILD_SHARE} %{BUILD_BIN} %{SYSTEMD} %{LOG}
 cp -r /usr/local/share/xbatd/* %{BUILD_SHARE}
-make -e BIN_DESTINATION=%{BUILD_BIN} SYS_DESTINATION=%{SYSTEMD} install 
+
+DESTDIR=%{buildroot} cmake --install build 
 
 %files
+%defattr(-,root,root,-)
 /usr/local/bin/xbatd
 /usr/local/share/xbatd
 /etc/systemd/system/xbatd.service
+%dir /var/log/xbatd
 
 %post
 systemctl daemon-reload
+systemctl enable xbatd.service
 
 %preun
-systemctl stop xbatd.service
+if [ $1 -eq 0 ]; then
+    systemctl stop xbatd.service
+    systemctl disable xbatd.service
+fi
 
 %postun
 systemctl daemon-reload
+if [ $1 -eq 0 ]; then
+    rm -rf /var/log/xbatd
+fi
+
+%clean
+rm -rf %{buildroot}
